server/requirements.txt
openenv-core
fastapi>=0.115.0
uvicorn>=0.24.0
openai>=1.40.0


